move test cases
